TOB-6147 Refaktorerer hvordan vi håndterer nytt skattekort i BestillingService * Oppretter bestilling for nytt fnr * Oppretter også bestilling for inneværende inntektsår når vi får nytt fnr fra PDL

+++ src/main/kotlin/no/nav/sokos/skattekort/skatteetaten/bestillskattekort/BestillSkattekortRequest.kt
@@ -66,7 +66,7 @@ fun bestillSkattekortRequest(
             ),
     )
 
-fun bestillOppdateringRequest(inntektsaar: Int): BestillSkattekortRequest =
+fun bestillOppdateringRequest(inntektsaar: Short): BestillSkattekortRequest =
     BestillSkattekortRequest(
         inntektsaar = inntektsaar.toString(),
         bestillingstype = "HENT_KUN_ENDRING",
